# Directly call cublas in awq_dequant