--- a/vllm/model_executor/layers/quantization/awq.py
+++ b/vllm/model_executor/layers/quantization/awq.py
@@ -154,12 +154,12 @@ class AWQLinearMethod(LinearMethodBase):
         # num_tokens >= threshold
         FP16_MATMUL_HEURISTIC_CONDITION = x.shape[:-1].numel() >= 256
 
-        if FP16_MATMUL_HEURISTIC_CONDITION:
-            out = ops.awq_dequantize(qweight, scales, qzeros, 0, 0, 0)
-            out = torch.matmul(reshaped_x, out)
-        else:
-            out = ops.awq_gemm(reshaped_x, qweight, scales, qzeros,
-                               pack_factor)
+        # if FP16_MATMUL_HEURISTIC_CONDITION:
+        #     out = ops.awq_dequantize(qweight, scales, qzeros, 0, 0, 0)
+        #     out = torch.matmul(reshaped_x, out)
+        # else:
+        out = ops.awq_gemm(reshaped_x, qweight, scales, qzeros,
+                            pack_factor)
         if bias is not None:
             out = out + bias
         return out.reshape(out_shape)