For long input sequences (>256), it performs better today with dequant + torch.matmul than the original awq_gemm. This PR changes to directly call cuBLAS instead of the torch.matmul.

Surprisingly it's much slower than the current main.

| input len | output len | main reqs/s | main tokens/s | PR reqs/s | PR tokens/s | Speed Up |
|-|-|-|-|-|-|-|
| 1024 | 64 | 8.61 | 9365.52 | 4.23 | 4597 | .49 |
| 512 | 64 | 14.24 | 8202.31 | 7.58 | 4368.18 | .53 |
| 256 | 64 | 20.78 | 6648.15 | 12.56 | 4018.48 | .60 |
| 128 | 128 | 15.75 | 4031.75 | 12.66 | 3239.88 | .80 |
| 64 | 64 | 32.44 | 4152.96 | 28.53 | 3651.43 | .88 |
| 32 | 64 | 35.44 | 3402.48 | 33.36 | 3202.88 | .94 |
| 32 | 32 | 65 | 4160.22 | 65.28 | 4177.73 | 1 |


This PR also refactored the awq_gemm which isn't the source of performance degrade. Observed similar results when directly adding the cublas gemm to the original awq_dequantize kernel.