@zcnrex 2024-02-20
@casper-hansen the result is surprisingly slow per our previous discussion. I also tried directly calling cuBLAS in the original `awq_dequantize` but achieved similar results. Could you provide any insight on how to speedup calling cuBLAS? Also could this be an opportunity to speed up GPTQ which uses the same `cublasHgemm`