@mgoin 2024-02-20
Perhaps torch.matmul is using the cutlass backend which can be more performant than cuBLAS?