@casper-hansen 2024-02-22
See this reference. If the number of tokens is larger than 32 (MAX_Q_GEMM_ROWS), then it dequantizes and calls `cublasHgemm` in the same kernel. As for the details, I'm not sure why the current implementation in this PR turns out to be slower - perhaps the matmul in torch is already plenty fast, or ExLlamaV2 is much faster at dequantizing, or there is some additional overhead that I have not been able to identify.

https://github.com/turboderp/exllamav2/blob/master/exllamav2/exllamav2_ext/cuda/q_gemm.cu#L122-L139